Moved the matching procedure out of the `Matching` class.
We define an "assignation pass" to be a procedure that given a list of
player returns a list of (player, activity) pairs for each player assign
to an activity at this pass.

This isolates the algorithmic part (the generation of an assignation
list) to the management part (managing availabilities, displaying
activities cast and players activities, etc.), where previously, the
`Matching` class did both.

The hospital-resident algorithm has been moved into a new class
`HospitalResident.py` in the `hospitalResident.py` file.

Finally, this will ease the definition of alternative matching
procedures. Indeed, it will just be necessary to define a
`assignation_pass` for some class inheriting the `Matching` base class.